requirements.txt
fastapi
google-cloud-storage
numpy
pydantic
scikit-image
tensorflow
tensorflow-hub
uvicorn